% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_dividends}
\alias{yf_dividends}
\title{Fetch dividend history}
\usage{
yf_dividends(ticker, start = Sys.Date() - 3650)
}
\arguments{
\item{ticker}{Ticker symbol}

\item{start}{Start date (default: 10 years ago)}
}
\value{
tibble: date, dividend, ticker
}
\description{
Fetch dividend history
}
