% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_options}
\alias{yf_options}
\title{Options chain}
\usage{
yf_options(ticker, expiration = NULL)
}
\arguments{
\item{ticker}{Ticker symbol}

\item{expiration}{Expiration date (Date or string). NULL = nearest.}
}
\value{
tibble: ticker, expiration, type (call/put), strike, last_price,
bid, ask, volume, open_interest, implied_volatility
}
\description{
Options chain
}
