% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_profile}
\alias{yf_profile}
\title{Company profile}
\usage{
yf_profile(ticker)
}
\arguments{
\item{ticker}{Ticker symbol}
}
\value{
tibble: ticker, name, sector, industry, employees, country,
city, website, description
}
\description{
Company profile
}
