% Generated by build script
\name{yf_recommendations}
\alias{yf_recommendations}
\title{Analyst recommendations and upgrades/downgrades}
\usage{
yf_recommendations()
}
\arguments{
\item{ticker}{Ticker symbol}
}
\value{tibble: ticker, firm, to_grade, from_grade, action, date}
\description{Analyst recommendations and upgrades/downgrades}
