% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_search}
\alias{yf_search}
\title{Search Yahoo Finance for tickers}
\usage{
yf_search(query)
}
\arguments{
\item{query}{Search string (e.g. "apple", "AAPL", "bitcoin")}
}
\value{
tibble: symbol, name, type, exchange
}
\description{
Search Yahoo Finance for tickers
}
