% Generated by build script
\name{yf_splits}
\alias{yf_splits}
\title{Fetch stock split history}
\usage{
yf_splits()
}
\arguments{
\item{ticker}{Ticker symbol}
\item{start}{Start date (default: 20 years ago)}
}
\value{tibble: date, numerator, denominator, ratio, ticker}
\description{Fetch stock split history}
