% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_statistics}
\alias{yf_statistics}
\title{Key statistics}
\usage{
yf_statistics(ticker)
}
\arguments{
\item{ticker}{Ticker symbol}
}
\value{
tibble (long format): ticker, metric, value
}
\description{
Key statistics
}
