% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yf.R
\name{yf_splits}
\alias{yf_splits}
\title{Fetch stock split history}
\usage{
yf_splits(ticker, start = Sys.Date() - 7300)
}
\arguments{
\item{ticker}{Ticker symbol}

\item{start}{Start date (default: 20 years ago)}
}
\value{
tibble: date, numerator, denominator, ratio, ticker
}
\description{
Fetch stock split history
}
